@@ -1,9 +1,10 @@
 module github.com/xukgo/log4z
 
-go 1.12
+go 1.16
 
 require (
-	go.uber.org/zap v1.14.1-0.20200305180913-9646482274d0
+	go.uber.org/atomic v1.10.0 // indirect
+	go.uber.org/multierr v1.9.0 // indirect
+	go.uber.org/zap v1.24.0
 	golang.org/x/lint v0.0.0-20200130185559-910be7a94367 // indirect
-	golang.org/x/tools v0.0.0-20200207001614-6fdc5776f4bb // indirect
 )
